Add manual file for APIs the metadata can't support
